Allow Sequence and Or to have only one child.
These constructs are meant for combining multiple patterns, so we
previously enforced that they must combine two or more subpatterns.

However, it makes sense for the a one-pattern sequence to be that
pattern itself.  Furthermore, code it will be easier to implement
pattern generation tools if we don't restrict Sequence and Or to take
multiple arguments.  So, they now only require one or more subpatterns.

(This problem comes about in the implementation of the recursive
descent parser for regular expressions.  The easiest way to enforce
order-of-operation rules involves assuming that every subpatterns of
any alternation is a concatenation.  If it turns out that that
subpattern is only a single, unconcatenated pattern, then it's easiest
to treat it as though it were a singleton concatenation.)