Do not create a state where an edge would suffice.
Previously, the system generated patterns for OneOf as though it were
an Or.  In turn, it created two new states per character in the OneOf
object.  That meant that the regex '.' was converted to an NFA with
over one hundred states!

Now, instead, a OneOf is represented as a start state and an end state
with many transitions from start to end-- one for each character.

This allows patterns like '.{1000}' to be transformed into NFAs much
more quickly (about 10x faster).