Use RewindSource, not str, for lexing.
RewindSource exposes a more natural interface.  The NFA reads as many
characters as it wishes, and adopts those characters.  It then tells
the RewindSource to act as though it had never read any more
characters.  At that point, the first character of the next token is
ready to be read.